updates for 26.04 + claude skill for this update

+++ python/tests/test_metrics.py
@@ -96,7 +96,7 @@ def test_multi_class_metrics(
     ).astype(np.float64)
 
     probabilities = np.random.rand(1000, num_classes)
-    probabilities[range(1000), list(pdf["label"].astype(np.integer))] = 2.0
+    probabilities[range(1000), list(pdf["label"].astype(int))] = 2.0
     probabilities = probabilities / np.sum(probabilities, axis=1).reshape(-1, 1)
 
     pdf["probabilities"] = list(probabilities)
